change request: move path filtering test into separate file, renamed ci_path_filters.py -> configure_ci_path_filters.py
