+++ build_tools/github_actions/tests/configure_ci_test.py
@@ -8,7 +8,6 @@ from unittest.mock import patch
 
 sys.path.insert(0, os.fspath(Path(__file__).parent.parent))
 import configure_ci
-from ci_path_filters import is_ci_run_required
 from benchmarks.benchmark_test_matrix import benchmark_matrix
 
 therock_test_runner_dict = {
@@ -56,56 +55,6 @@ class ConfigureCITest(unittest.TestCase):
             self.assertFalse(
                 any(entry.get("expect_failure") for entry in target_output)
             )
-
-    ###########################################################################
-    # Tests for is_ci_run_required
-
-    def test_run_ci_if_source_file_edited(self):
-        paths = ["source_file.h"]
-        run_ci = is_ci_run_required(paths)
-        self.assertTrue(run_ci)
-
-    def test_dont_run_ci_if_only_markdown_files_edited(self):
-        paths = ["README.md", "build_tools/README.md"]
-        run_ci = is_ci_run_required(paths)
-        self.assertFalse(run_ci)
-
-    def test_dont_run_ci_if_only_external_builds_edited(self):
-        paths = ["external-builds/pytorch/CMakeLists.txt"]
-        run_ci = is_ci_run_required(paths)
-        self.assertFalse(run_ci)
-
-    def test_dont_run_ci_if_only_external_builds_edited(self):
-        paths = ["experimental/file.h"]
-        run_ci = is_ci_run_required(paths)
-        self.assertFalse(run_ci)
-
-    def test_run_ci_if_related_workflow_file_edited(self):
-        paths = [".github/workflows/ci.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertTrue(run_ci)
-
-        paths = [".github/workflows/build_portable_linux_artifacts.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertTrue(run_ci)
-
-        paths = [".github/workflows/build_artifact.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertTrue(run_ci)
-
-    def test_dont_run_ci_if_unrelated_workflow_file_edited(self):
-        paths = [".github/workflows/pre-commit.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertFalse(run_ci)
-
-        paths = [".github/workflows/test_jax_dockerfile.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertFalse(run_ci)
-
-    def test_run_ci_if_source_file_and_unrelated_workflow_file_edited(self):
-        paths = ["source_file.h", ".github/workflows/pre-commit.yml"]
-        run_ci = is_ci_run_required(paths)
-        self.assertTrue(run_ci)
 
     ###########################################################################
     # Tests for matrix_generator and helper functions
